2.5 Creating scripts
Since I already created a script derived from "MonoBehaviour" without checking if this is a step in the later stages, this step has been already handled too. In this commit, I put that script in a scripts folder as it was asked in the pdf. I've commit the "Tag Manager" script in its own commit which can be seen in the history.