2.9 Adding lighting & 2.10 Modifying the Camera
I added a point light and a spot light. I reset the size of a side wall, as it wasn't required to be that size now. Instead, I expanded another wall to reflect the shadows of the spot light. I lowered the directional light to better reflect the shadows.

I know we need to commit different steps one at a time but I already adjusted the camera to better reflect the scene like I did in previous steps. So this commit is for 2 steps combined.