2.14 (Optional) Making the camera follow the player
I took the longer way. Simply attaching the camera to the player would've worked but I wanted to add functionality like a third person game where player is able to look around without changing the rotation of the players mesh. I also wanted it to be detached because I'm spinning the mesh via C and Z buttons as I've implemented in the previous commits and I didn't want the camera to spin around with it.