Ice Prefab
Created the main prefab for the Ice segment of the game. The Ice is extremely slippery, has increased speed and force values.

As in the lava segment, might add some more little meshes later as the maze becomes a whole. I'll probably add some angle changes and color changes on crystals in the wall later.

I'll probably switch "enter" and "exit" collision checks to transform checks. The current form is open to exploitation.