Invisible Prefab
Created the main prefab for the Invisible segment of the game. As you progress through the segment, breadcrumbs appear so that you can trace back your steps if needed. I'll add an additional navigation system in later steps.

As in the lava segment, might add some more little meshes later as the maze becomes a whole. I'll probably add some angle changes and color changes on crystals in the wall later.

I'll probably switch "enter" and "exit" collision checks to transform checks. The current form is open to exploitation.